chore(docs): remove Disqus
We found that it's better for developers using Angular to contribute to the
docs via a PR on Github rather than a comment at the bottom of the docs page.
Similarly, we found that StackOverflow or our IRC channel (#angularjs on
freenode) were better outlets for questions.

We think that removing the comments and directing users towards these channels
will ultimately be more productive.